requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
backoff>=2.2.1
tabulate>=0.9.0

# Model providers (install only what you need)
openai>=1.40.0
anthropic>=0.34.0
google-genai>=0.3.0

# Evaluation metrics
rouge-score>=0.1.2
scikit-learn>=1.3.0
scipy>=1.11.0

# Optional: local Hugging Face inference (Qwen, etc.)
# transformers>=4.44.0
# torch>=2.1.0
# accelerate>=0.30.0

# Optional: embedding-based semantic match fallback (no OpenAI)
# sentence-transformers>=2.7.0
